requirements.txt
-i https://pypi.org/simple
blinker==1.8.2; python_version >= '3.8'
brotli==1.1.0; platform_python_implementation != 'PyPy'
click==8.1.7; python_version >= '3.7'
flask==3.0.3; python_version >= '3.8'
flask-compress==1.15
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.1.1; python_version < '3.13' and (platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32'))))))
gunicorn==23.0.0; python_version >= '3.7'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
packaging==24.1; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
sqlalchemy==2.0.35; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
werkzeug==3.0.4; python_version >= '3.8'
zstandard==0.23.0; platform_python_implementation != 'PyPy'
